fix: tier 릴레이션 속성 이름 변경

+++ Phase_4/server/src/main/java/fift/server/service/gifticon/GifticonService.java
@@ -37,7 +37,7 @@ public class GifticonService {
         if (currentMoney >= productPrice) {
             double updatedMoney = currentMoney - productPrice;
             if (updatedMoney >= 0) {
-                customer.setMoney((long) updatedMoney);
+                customer.setMoney((double) updatedMoney);
                 customer.setAmount(customer.getAmount() + productPrice);
                 customerRepository.save(customer);
